Refactored how we handle references to entities
Instead of a pointer which can get freed without our knowledge, we now use an index + generation pair that we call EntityId. Entities are never destroyed, instead they are set as available for re-use. If we hold an EntityId to an entity that no longer exists or has been re-used, trying to get it will simply return null so we can't access the removed entity.